+++ include/coyote/strategies/random_strategy.h
@@ -32,7 +32,7 @@ namespace coyote
 		RandomStrategy& operator=(RandomStrategy const&) = delete;
 
 		// Returns the next operation.
-		size_t next_operation(Operations& operations, size_t current);
+		int next_operation(Operations& operations, size_t current);
 
 		// Returns the next boolean choice.
 		bool next_boolean();
